frontend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
google-generativeai
numpy
Pillow
PyJWT==2.8.0
bcrypt==4.1.2
SQLAlchemy==2.0.25
